将所有float类型替换为double类型，同时改进测试函数

+++ include/SpectralIndices_interface.h
@@ -11,6 +11,6 @@ using namespace std;
 /// @param spectralindices_index 光谱指数索引（详情见SpectralIndices_math.h）
 /// @param result 存放结果的二维数组指针
 /// @return 成功返回1，失败返回0
-int spectral_index_calculation(int ***raw_data, int x, int y, int z, float *value_array_of_all_band,
-                               SpectralIndices spectralindices_index, float **result);
+int spectral_index_calculation(int ***raw_data, int x, int y, int z, double *value_array_of_all_band,
+                               SpectralIndices spectralindices_index, double **result);
